Use uint32_t for indexes

+++ src/base/base.h
@@ -147,7 +147,7 @@ template <class T> struct LinkedNode {
 
 template <class T> struct LinkedList {
   LinkedNode<T> *head = nullptr;
-  size_t size = 0;
+  uint32_t size = 0;
 
   T *emplace_front(BumpArena &arena) {
     LinkedNode<T> *node = arena.alloc<LinkedNode<T>>();
@@ -160,16 +160,16 @@ template <class T> struct LinkedList {
 
 template <class T> struct Array {
   T *data;
-  size_t size;
+  uint32_t size;
 
-  static Array<T> create(BumpArena &arena, size_t with_size) {
+  static Array<T> create(BumpArena &arena, uint32_t with_size) {
     T *result =
         static_cast<T *>(arena.alloc_raw(with_size * sizeof(T), alignof(T)));
     return Array<T>{result, with_size};
   }
 
   static Array<T> copy_from(BumpArena &arena, const T *from,
-                            const size_t size) {
+                            const uint32_t size) {
     Array<T> dst = create(arena, size);
     if (size > 0) {
       memcpy(dst.data, from, size * sizeof(T));
@@ -186,14 +186,14 @@ template <class T> struct Array {
 
 template <class T> struct GrowingArray {
   Array<T> inner;
-  size_t cur_size;
+  uint32_t cur_size;
 
   T *emplace_back(BumpArena &arena) {
-    size_t wasted = 0;
+    uint32_t wasted = 0;
     return emplace_back(arena, wasted);
   }
 
-  T *emplace_back(BumpArena &arena, size_t &wasted_bytes) {
+  T *emplace_back(BumpArena &arena, uint32_t &wasted_bytes) {
     if (cur_size >= inner.size) {
       wasted_bytes += inner.byte_size();
       realloc(arena);
@@ -202,13 +202,13 @@ template <class T> struct GrowingArray {
   }
 
   void realloc(BumpArena &arena) {
-    size_t new_size = std::max(static_cast<size_t>(4), cur_size * 2);
+    uint32_t new_size = std::max(4u, cur_size * 2);
     Array<T> new_inner = Array<T>::create(arena, new_size);
     if (inner.data) memcpy(new_inner.data, inner.data, cur_size * sizeof(T));
     inner = new_inner;
   }
 
-  void shrink_to(size_t size) {
+  void shrink_to(uint32_t size) {
     if (size >= cur_size) return;
     memset(inner.data + size, 0, (cur_size - size) * sizeof(T));
     cur_size = size;
@@ -216,7 +216,7 @@ template <class T> struct GrowingArray {
 
   T *data() { return inner.data; }
   const T *data() const { return inner.data; }
-  size_t size() const { return cur_size; }
+  uint32_t size() const { return cur_size; }
   size_t total_byte_size() const { return inner.size * sizeof(T); }
 
   T last_or(T def) const {
@@ -230,11 +230,11 @@ template <class T> struct GrowingArray {
 };
 
 template <class T, class F>
-size_t lower_bound(const size_t size, F get_val, const T val) {
-  size_t left = 0;
-  size_t right = size;
+uint32_t lower_bound(const uint32_t size, F get_val, const T val) {
+  uint32_t left = 0;
+  uint32_t right = size;
   while (right - left > 1) {
-    const size_t mid = (left + right) / 2;
+    const uint32_t mid = (left + right) / 2;
     const T &mid_val = get_val(mid);
     if (mid_val <= val) {
       left = mid;
@@ -246,10 +246,10 @@ size_t lower_bound(const size_t size, F get_val, const T val) {
 }
 
 template <class T, class F>
-size_t bin_search_exact(const size_t size, F get_val, const T val) {
-  const size_t result = lower_bound(size, get_val, val);
+uint32_t bin_search_exact(const uint32_t size, F get_val, const T val) {
+  const uint32_t result = lower_bound(size, get_val, val);
   if (result >= size || get_val(result) != val) {
-    return SIZE_MAX;
+    return UINT32_MAX;
   }
   return result;
 }
