+++ src/base/string.h
@@ -4,11 +4,11 @@
 
 struct String {
   const char *data;
-  size_t len;
+  uint32_t len;
 
   static String copy_from(BumpArena &arena, const char *from,
-                          const size_t len) {
-    const size_t size = len + 1;
+                          const uint32_t len) {
+    const uint32_t size = len + 1;
     char *dst = static_cast<char *>(arena.alloc_raw(size, 1));
     dst[len] = '\0';
     memcpy(dst, from, len);
@@ -16,12 +16,12 @@ struct String {
   }
 
   static String copy_from(BumpArena &arena, const char *from) {
-    return copy_from(arena, from, strlen(from));
+    return copy_from(arena, from, static_cast<uint32_t>(strlen(from)));
   }
 
   static String copy_from(BumpArena &arena, const String &from) {
     return copy_from(arena, from.data, from.len);
   }
 
-  size_t byte_size() const { return len + 1; }
+  uint32_t byte_size() const { return len + 1; }
 };
