+++ src/views/environ_viewer.h
@@ -40,7 +40,7 @@ struct EnvironViewerWindow {
 struct EnvironViewerState {
   GrowingArray<EnvironViewerWindow> windows;
   BumpArena cur_arena;
-  size_t updates_since_last_cleanup;
+  uint32_t updates_since_last_cleanup;
 };
 
 struct FrameContext;
